moved binary-search functions from prune_verts into core
