+++ python/_prune_unused_bones.py
@@ -22,12 +22,12 @@
 try:
 	from . import nuthouse01_core as core
 	from . import nuthouse01_pmx_parser as pmxlib
-	from ._prune_unused_vertices import newval_from_range_map, delme_list_to_rangemap, binary_search_isin
+	from ._prune_unused_vertices import newval_from_range_map, delme_list_to_rangemap
 except ImportError as eee:
 	try:
 		import nuthouse01_core as core
 		import nuthouse01_pmx_parser as pmxlib
-		from _prune_unused_vertices import newval_from_range_map, delme_list_to_rangemap, binary_search_isin
+		from _prune_unused_vertices import newval_from_range_map, delme_list_to_rangemap
 	except ImportError as eee:
 		print(eee.__class__.__name__, eee)
 		print("ERROR: failed to import some of the necessary files, all my scripts must be together in the same folder!")
@@ -35,7 +35,7 @@ except ImportError as eee:
 		input()
 		exit()
 		core = pmxlib = None
-		newval_from_range_map = delme_list_to_rangemap = binary_search_isin = None
+		newval_from_range_map = delme_list_to_rangemap = None
 
 
 # when debug=True, disable the catchall try-except block. this means the full stack trace gets printed when it crashes,
@@ -282,7 +282,7 @@ def apply_bone_remapping(pmx, bone_dellist, bone_shiftmap):
 		i = 0
 		while i < len(morph[4]):
 			# if the bone being manipulated is in the list of bones being deleted, delete it here too. otherwise remap.
-			if binary_search_isin(morph[4][i][0], bone_dellist):
+			if core.binary_search_isin(morph[4][i][0], bone_dellist):
 				morph[4].pop(i)
 			else:
 				morph[4][i][0] = newval_from_range_map(morph[4][i][0], bone_shiftmap)
@@ -300,7 +300,7 @@ def apply_bone_remapping(pmx, bone_dellist, bone_shiftmap):
 				i += 1
 			else:
 				# if this is one of the bones being deleted, delete it here too. otherwise remap.
-				if binary_search_isin(item[1], bone_dellist):
+				if core.binary_search_isin(item[1], bone_dellist):
 					frame[3].pop(i)
 				else:
 					item[1] = newval_from_range_map(item[1], bone_shiftmap)
@@ -319,7 +319,7 @@ def apply_bone_remapping(pmx, bone_dellist, bone_shiftmap):
 	for d, bone in enumerate(pmx[5]):
 		# point-at link:
 		if bone[12]:
-			if binary_search_isin(bone[13][0], bone_dellist):
+			if core.binary_search_isin(bone[13][0], bone_dellist):
 				# if pointing at a bone that will be deleted, instead change to offset with offset 0,0,0
 				bone[12] = 0
 				bone[13] = [0, 0, 0]
@@ -331,7 +331,7 @@ def apply_bone_remapping(pmx, bone_dellist, bone_shiftmap):
 		bone[5] = newval_from_range_map(bone[5], bone_shiftmap)
 		# partial append:
 		if (bone[14] or bone[15]) and bone[16][1] != 0:
-			if binary_search_isin(bone[16][0], bone_dellist):
+			if core.binary_search_isin(bone[16][0], bone_dellist):
 				# if a bone is getting partial append from a bone getting deleted, break that relationship
 				bone[14] = 0
 				bone[15] = 0
